Solved: Median of Two Sorted Arrays (#4)

+++ problems/0004-median-of-two-sorted-arrays/solution.cpp
@@ -5,34 +5,34 @@ public:
         int cnt = 0;
         int n1 = nums1.size(), n2 = nums2.size();
         int n = (n1 + n2);
-        int ind1 = n/2 - 1, ind2 = n/2;
+        double ind1 = n/2 - 1, ind2 = n/2;
         double indlen1 = -1, indlen2 = -1;
         while(i < n1 && j < n2){
             if((indlen1 != -1) && (indlen2 != -1)) break;
             if(nums1[i] < nums2[j]){
-                if(cnt == ind1) indlen1 = nums1[i];
-                if(cnt == ind2) indlen2 = nums2[i];
+                if(cnt == ind1) (double)indlen1 = nums1[i];
+                if(cnt == ind2) (double)indlen2 = nums2[i];
                 cnt++;
                 i++;
             }
             else{
-                if(cnt == ind1) indlen1 = nums1[j];
-                if(cnt == ind2) indlen2 = nums2[j];
+                if(cnt == ind1) (double)indlen1 = nums1[j];
+                if(cnt == ind2) (double)indlen2 = nums2[j];
                 cnt++;
                 j++;               
             }
         }
         while(i < n1){
                 if((indlen1 != -1) && (indlen2 != -1)) break;
-                if(cnt == ind1) indlen1 = nums1[i];
-                if(cnt == ind2) indlen2 = nums2[i];
+                if(cnt == ind1) (double)indlen1 = nums1[i];
+                if(cnt == ind2) (double)indlen2 = nums2[i];
                 cnt++;
                 i++;            
         }
         while(j < n2){
                 if((indlen1 != -1) && (indlen2 != -1)) break;
-                if(cnt == ind1) indlen1 = nums1[j];
-                if(cnt == ind2) indlen2 = nums2[j];
+                if(cnt == ind1) (double)indlen1 = nums1[j];
+                if(cnt == ind2) (double)indlen2 = nums2[j];
                 cnt++;
                 j++;            
         }
